GenAi-Prototype/requirements.txt
streamlit
pandas
numpy
plotly
altair
matplotlib
python-dotenv
openai
google-genai
google-generativeai